requirements.txt
GDAL==1.11.0
Pillow==2.5.2
arrow==0.4.4
beautifulsoup4==4.3.2
gunicorn==19.1.1
oauthlib==0.6.3
python-dateutil==2.2
pytz==2014.4
requests==2.1.0
requests-oauthlib==0.4.0
six==1.7.3
twython==3.1.2
wsgiref==0.1.2
